perf(plugin): speed up inference middleware imports
