+++ plugins/nemo-switchyard/tests/test_integration.py
@@ -20,6 +20,8 @@ from nemo_platform_plugin.inference_middleware import (
     InferenceMiddlewareContext,
     InferenceRequest,
     InferenceResponse,
+)
+from nemo_platform_plugin.inference_middleware_models import (
     MiddlewareCall,
     VirtualModel,
     VirtualModelInferenceConfig,
